Time: 7 ms (13.88%), Space: 44.5 MB (11.05%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -20,7 +20,7 @@ class Solution {
 
         
 
-        return reverse == (long) x ;
+        return reverse == x ;
         
     }
 }
